Adds a check for equilibrium file when running GEM for every folder in the original repository

+++ run_gem_standalone_check.sh
@@ -27,8 +27,16 @@ for f in $(find . -type f -name *coreprof*cpo); do
     cp ${f}   ${orig_repo_loc}/${new_run_dir}/${d_path}/${d_name}/gem_coreprof_in.cpo
 
     # copy xml, xsd, sh, cpo-s and exectuables to every new directory
-    
-    cp ${orig_repo_loc}/workflows/AUG_28906_6/ets_equilibrium_in.cpo   ${orig_repo_loc}/${new_run_dir}/${d_path}/${d_name}/gem_equilibrium_in.cpo
+
+    # if the folder contains and equilibrium file, copy it, else get a default one
+    if [ -e ./${d_path}/*equilibrium*.cpo ] 
+    then
+      echo "using local equilibrium in "${d_path}/${d_name}
+      cp ./${d_path}/*equilibrium*.cpo  ${orig_repo_loc}/${new_run_dir}/${d_path}/${d_name}/gem_equilibrium_in.cpo
+    else
+      echo "using default equilibrium in "${d_path}/${d_name}     
+      cp ${orig_repo_loc}/workflows/AUG_28906_6/ets_equilibrium_in.cpo   ${orig_repo_loc}/${new_run_dir}/${d_path}/${d_name}/gem_equilibrium_in.cpo
+    fi
 
     cp ${orig_repo_loc}/workflows/gem_1n_8ft.xml   ${orig_repo_loc}/${new_run_dir}/${d_path}/${d_name}/gem.xml
     cp ${orig_repo_loc}/workflows/gem.xsd   ${orig_repo_loc}/${new_run_dir}/${d_path}/${d_name}/
@@ -44,5 +52,7 @@ for f in $(find . -type f -name *coreprof*cpo); do
 
 done
 
+# display the ion heat fluxes at time step #100 for all the runs
 
+grep -iRn ti_transp%flux --include="gem_coretransp_0100.cpo" -A 5
